requirements.txt
bleach==3.0.2
certifi==2018.10.15
chardet==3.0.4
Django>=1.11.0
django-nose==1.4.6
djangorestframework
docutils==0.14
flake8==3.6.0
idna==2.7
mccabe==0.6.1
nose==1.3.7
pkginfo==1.4.2
pycodestyle==2.4.0
pyflakes==2.0.0
Pygments==2.2.0
pytz==2018.7
readme-renderer==24.0
requests==2.20.1
requests-toolbelt==0.8.0
schema==0.6.8
six==1.11.0
tqdm==4.28.1
twine==1.12.1
urllib3==1.24.1
webencodings==0.5.1
